pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "contamination"
version = "1.0.0"
authors = [
  { name="Anonymous", email="anonymous@anonymous.com" },
]
description = "Contamination Package"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "accelerate >= 0.25.0",
  "aiohttp >= 3.9.1",
  "aiosignal >= 1.3.1",
  "datasets >= 2.16.0",
  "ipykernel>=6.26.0",
  "loguru >= 0.7.2",
  "matplotlib >= 3.8.2",
  "numpy >= 1.24.3",
  "openai >= 1.6.1",
  "pandas >= 2.1.4",
  "peft >= 0.7.1",
  "python-dotenv >= 1.0.0",
  "rouge >= 1.0.1",
  "rouge-score >= 0.1.2",
  "sacrebleu >= 2.4.0",
  "scikit-learn >= 1.3.2",
  "seaborn >= 0.13.0",
  "sentencepiece >= 0.1.99",
  "tokenizers >= 0.15.0",
  "torch >= 2.1.2",
  "tqdm >= 4.66.1",
  "transformers >= 4.36.2",
  "triton >= 2.1.0",
  "trl >= 0.7.6"
]